requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
